requirements.txt
torch
transformers
Pillow
streamlit
